+++ CHANGELOG.md
@@ -8,8 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
-- ModelService.GetComputeTaskInputModels, use ComputeTaskAPI.GetInputAssets instead
+- (BREAKING) ModelService.GetComputeTaskInputModels, use ComputeTaskAPI.GetInputAssets instead
 - Test task rank special case: rank is not inherited from parent task anymore
+- (BREAKING) QueryModels from model service: it was unused and model category will soon be deprecated
 
 ### Deprecated